Consumers/requirements.txt
anyio==3.7.1
certifi==2023.5.7
charset-normalizer==3.2.0
confluent-kafka==2.2.0
exceptiongroup==1.1.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
joblib==1.2.0
ksqldb==1.0.1
numpy==1.25.1
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.0.3
xgboost==1.7.6
